+++ frontend/slices/notion-shell/index.ts
@@ -42,6 +42,8 @@ export {
 } from "./components/SortableBlockList";
 export { PageActionsMenu, type PageActionsMenuProps } from "./components/PageActionsMenu";
 export { PageLayoutSection, type PageLayoutSectionProps, type PageFont } from "./components/PageLayoutSection";
+export { PageBreadcrumbs, type PageBreadcrumbsProps, type Crumb } from "./components/PageBreadcrumbs";
+export { Subpages, SubpageIcon, type SubpagesProps, type SubpageItem } from "./components/Subpages";
 export { InlineFormatToolbar } from "./components/InlineFormatToolbar";
 export { BlockColorPicker } from "./components/BlockColorPicker";
 export { BLOCK_COLORS, blockColorClass, type BlockColor } from "./lib/blockColors";